feat: move TimetableAdapter to separate project
