+++ ExamSchedule.sln
@@ -8,6 +8,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ScheduleParserCLI", "Schedu
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExamScheduleTests", "ExamScheduleTests\ExamScheduleTests.csproj", "{C2166CD1-9377-4F01-A67B-C2466FA68B1B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TimetableAdapter", "TimetableAdapter\TimetableAdapter.csproj", "{8B2348AF-A29E-400F-979A-F3D5D41E5AAA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -30,5 +32,9 @@ Global
 		{C2166CD1-9377-4F01-A67B-C2466FA68B1B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C2166CD1-9377-4F01-A67B-C2466FA68B1B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C2166CD1-9377-4F01-A67B-C2466FA68B1B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8B2348AF-A29E-400F-979A-F3D5D41E5AAA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8B2348AF-A29E-400F-979A-F3D5D41E5AAA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8B2348AF-A29E-400F-979A-F3D5D41E5AAA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8B2348AF-A29E-400F-979A-F3D5D41E5AAA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
