+++ frontend/package.json
@@ -34,6 +34,7 @@
     "eslint": "^8.55.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.5",
+    "playwright": "^1.44.1",
     "typescript": "^5.2.2",
     "vite": "^5.0.8"
   }
